+++ CHANGELOG.md
@@ -13,6 +13,7 @@ ChangeLog
 - IMPROVEMENT: add Pane and Pillar to control UI and rendering
 - IMPROVEMENT: add Mask as a new fullscreen element in the front layer
 - IMPROVEMENT: add Media for loading image and video files into the Displays
+- IMPROVEMENT: add support for preset locking/unlocking
 - BUGFIX: fix preset filters by adding Plate and Hologram
 - BUGFIX: fix rendering of decal
 - UPDATE: update dependencies